host-tools: change prerequisites for easier/smaller/faster builds from containers
- sort out additional "groups" from prerequisites file for better overview,
  adds categories "compressor", "decompressor" and "checksum"
- change the check_prerequisites script to accept these new groups

It's much easier (or at least much faster) to install the needed binaries
for cryptographical checksums, data decompression and data compression from
the package repositories of a Linux distribution - especially when running
the whole thing from a container and you've even to install many of the
prerequisites. Adding some more packages to install isn't really a hard task.