host-tools: remove unnecessary dependencies for some host tools
A 'tar' command was a prerequisite already and if the other applets from
the BusyBox for the build host are installed from their "native" packages,
there's no reason to depend on 'busybox-host' anymore.

The 'tar-host' target from 'include/make/020-helper-functions.mk' is
equally obsolete and therefor the dependency on the whole
UNPACK_TARBALL_PREREQUISITES value is unnecessary for these five
packages now.

Now we can create the SquashFS tools for V3 and V4 format without
further overhead, because only needed packages get built. So another
build target 'squashfs-host-tools' was added as a "single" target for
V3 und V4 SquashFS tools.